requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
google-cloud-translate==1.3.1
Wand==0.4.4
PyPDF2==1.26.0
google-cloud-vision==0.32.0
Flask-Cors==3.0.6
